+++ eslint.config.mjs
@@ -6,7 +6,15 @@ export default tseslint.config(
   eslint.configs.recommended,
   ...tseslint.configs.recommended,
   prettier,
-  { ignores: ['dist/**', 'node_modules/**', 'coverage/**'] },
+  {
+    ignores: [
+      'dist/**',
+      '**/dist/**',
+      'node_modules/**',
+      'coverage/**',
+      'examples/**',
+    ],
+  },
   {
     files: ['src/**/*.ts', 'tests/**/*.ts'],
     rules: {